Move structs related to error handling into `openai_bootstrap`
so that they can be used in `openai_proc_macros`
