+++ openai_bootstrap/Cargo.toml
@@ -8,3 +8,7 @@ license.workspace = true
 publish = false
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies.serde]
+version = "1.0"
+features = ["derive"]
