+++ openai_bootstrap/src/lib.rs
@@ -1,3 +1,5 @@
+use serde::Deserialize;
+
 pub const BASE_URL: &str = "https://api.openai.com/v1/";
 
 #[macro_export]
@@ -14,4 +16,33 @@ macro_rules! authorization {
 
         $request.header(AUTHORIZATION, format!("Bearer {token}"))
     }};
+}
+
+#[derive(Deserialize)]
+pub struct ErrorResponse {
+    pub error: OpenAiError,
+}
+
+#[derive(Deserialize, Debug)]
+pub struct OpenAiError {
+    pub message: String,
+    #[serde(rename = "type")]
+    pub error_type: String,
+    pub param: Option<String>,
+    pub code: Option<String>,
+}
+
+impl std::fmt::Display for OpenAiError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.message)
+    }
+}
+
+impl std::error::Error for OpenAiError {}
+
+#[derive(Deserialize)]
+#[serde(untagged)]
+pub enum ApiResponse<T> {
+    Ok(T),
+    Err(ErrorResponse),
 }
